--- a/pkg/utils/utils.go
+++ b/pkg/utils/utils.go
@@ -84,16 +84,9 @@ func RandString(rnd *rand.Rand, ln int) string {
 	return string(out[:ln])
 }
 
-var testUUIDFromTime bool
-
 func UUIDFromTime(rnd *rand.Rand) string {
-	if testUUIDFromTime {
+	if UnderTest {
 		return gocql.TimeUUIDWith(rnd.Int63(), 0, []byte("127.0.0.1")).String()
 	}
 	return gocql.UUIDFromTime(RandTime(rnd)).String()
-}
-
-func SetTestUUIDFromTime() {
-	// Makes TYPE_ASCII, TYPE_TEXT, TYPE_VARCHAR, TYPE_BLOB predictable
-	testUUIDFromTime = true
 }