Every time we generate an operation we recreate statemet. 
Since some of the statement content could be cached we need to do that in sake of speeding things up.