@dkropachev 2023-05-30
It shows 40% performance improvement on statements generations
Results before:
```
goos: linux
goarch: amd64
pkg: github.com/scylladb/gemini/pkg/generators
cpu: 12th Gen Intel(R) Core(TM) i9-12900HK
BenchmarkGenInsertStmt
BenchmarkGenInsertStmt/pk1_ck0_col0
BenchmarkGenInsertStmt/pk1_ck0_col0-20         	 3995550	       300.0 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1
BenchmarkGenInsertStmt/pk1_ck1_col1-20         	 1327195	       882.0 ns/op
BenchmarkGenInsertStmt/pk3_ck3_col5
BenchmarkGenInsertStmt/pk3_ck3_col5-20         	  481899	      2402 ns/op
BenchmarkGenInsertStmt/pkAll_ckAll_colAll
BenchmarkGenInsertStmt/pkAll_ckAll_colAll-20   	  113185	     10026 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1cr
BenchmarkGenInsertStmt/pk1_ck1_col1cr-20       	 1548429	       779.0 ns/op
BenchmarkGenInsertStmt/pk3_ck3_col3cr
BenchmarkGenInsertStmt/pk3_ck3_col3cr-20       	  604590	      2144 ns/op
BenchmarkGenInsertStmt/pk1_ck0_col0_lwt
BenchmarkGenInsertStmt/pk1_ck0_col0_lwt-20     	 3259177	       325.5 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1_lwt
BenchmarkGenInsertStmt/pk1_ck1_col1_lwt-20     	 1412473	       860.0 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1cr_lwt
BenchmarkGenInsertStmt/pk1_ck1_col1cr_lwt-20   	 1597159	       750.4 ns/op
BenchmarkGenInsertStmt/pkAll_ckAll_colAll_lwt
BenchmarkGenInsertStmt/pkAll_ckAll_colAll_lwt-20         	  119487	      9259 ns/op
BenchmarkGenUpdateStmt
BenchmarkGenUpdateStmt/pk1_ck0_col0
BenchmarkGenUpdateStmt/pk1_ck0_col0-20                   	 3132351	       387.3 ns/op
BenchmarkGenUpdateStmt/pk1_ck1_col1
BenchmarkGenUpdateStmt/pk1_ck1_col1-20                   	 1248636	       957.2 ns/op
BenchmarkGenUpdateStmt/pk3_ck3_col5
BenchmarkGenUpdateStmt/pk3_ck3_col5-20                   	  381147	      3043 ns/op
BenchmarkGenUpdateStmt/pkAll_ckAll_colAll
BenchmarkGenUpdateStmt/pkAll_ckAll_colAll-20             	   95241	     13774 ns/op
BenchmarkGenUpdateStmt/pk1_ck1_col1cr
BenchmarkGenUpdateStmt/pk1_ck1_col1cr-20                 	 1470493	       753.9 ns/op
BenchmarkGenUpdateStmt/pk3_ck3_col3cr
BenchmarkGenUpdateStmt/pk3_ck3_col3cr-20                 	  519759	      2230 ns/op
BenchmarkGenDeleteRows
BenchmarkGenDeleteRows/pk1_ck0_col1
BenchmarkGenDeleteRows/pk1_ck0_col1-20                   	 3350074	       335.2 ns/op
BenchmarkGenDeleteRows/pk1_ck1_col1
BenchmarkGenDeleteRows/pk1_ck1_col1-20                   	 1324776	       898.7 ns/op
BenchmarkGenDeleteRows/pk3_ck3_col5
BenchmarkGenDeleteRows/pk3_ck3_col5-20                   	  786490	      1548 ns/op
BenchmarkGenDeleteRows/pkAll_ckAll_colAll
BenchmarkGenDeleteRows/pkAll_ckAll_colAll-20             	  249268	      5566 ns/op
BenchmarkGenDeleteRows/pk1_ck1_col1cr
BenchmarkGenDeleteRows/pk1_ck1_col1cr-20                 	 1375662	       889.5 ns/op
BenchmarkGenDeleteRows/pk3_ck3_col3cr
BenchmarkGenDeleteRows/pk3_ck3_col3cr-20                 	  768820	      1538 ns/op
PASS
```

After:
```
goos: linux
goarch: amd64
pkg: github.com/scylladb/gemini/pkg/generators
cpu: 12th Gen Intel(R) Core(TM) i9-12900HK
BenchmarkGenInsertStmt
BenchmarkGenInsertStmt/pk1_ck0_col0
BenchmarkGenInsertStmt/pk1_ck0_col0-20         	 6686106	       171.3 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1
BenchmarkGenInsertStmt/pk1_ck1_col1-20         	 2498922	       473.4 ns/op
BenchmarkGenInsertStmt/pk3_ck3_col5
BenchmarkGenInsertStmt/pk3_ck3_col5-20         	  837535	      1479 ns/op
BenchmarkGenInsertStmt/pkAll_ckAll_colAll
BenchmarkGenInsertStmt/pkAll_ckAll_colAll-20   	  162674	      6571 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1cr
BenchmarkGenInsertStmt/pk1_ck1_col1cr-20       	 3058292	       378.0 ns/op
BenchmarkGenInsertStmt/pk3_ck3_col3cr
BenchmarkGenInsertStmt/pk3_ck3_col3cr-20       	  972315	      1180 ns/op
BenchmarkGenInsertStmt/pk1_ck0_col0_lwt
BenchmarkGenInsertStmt/pk1_ck0_col0_lwt-20     	 6726502	       174.7 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1_lwt
BenchmarkGenInsertStmt/pk1_ck1_col1_lwt-20     	 2522307	       486.5 ns/op
BenchmarkGenInsertStmt/pk1_ck1_col1cr_lwt
BenchmarkGenInsertStmt/pk1_ck1_col1cr_lwt-20   	 3141670	       385.5 ns/op
BenchmarkGenInsertStmt/pkAll_ckAll_colAll_lwt
BenchmarkGenInsertStmt/pkAll_ckAll_colAll_lwt-20         	  171283	      6914 ns/op
BenchmarkGenUpdateStmt
BenchmarkGenUpdateStmt/pk1_ck0_col0
BenchmarkGenUpdateStmt/pk1_ck0_col0-20                   	 6542786	       177.5 ns/op
BenchmarkGenUpdateStmt/pk1_ck1_col1
BenchmarkGenUpdateStmt/pk1_ck1_col1-20                   	 2389880	       502.7 ns/op
BenchmarkGenUpdateStmt/pk3_ck3_col5
BenchmarkGenUpdateStmt/pk3_ck3_col5-20                   	  764728	      1547 ns/op
BenchmarkGenUpdateStmt/pkAll_ckAll_colAll
BenchmarkGenUpdateStmt/pkAll_ckAll_colAll-20             	  170385	      7060 ns/op
BenchmarkGenUpdateStmt/pk1_ck1_col1cr
BenchmarkGenUpdateStmt/pk1_ck1_col1cr-20                 	 3323509	       371.8 ns/op
BenchmarkGenUpdateStmt/pk3_ck3_col3cr
BenchmarkGenUpdateStmt/pk3_ck3_col3cr-20                 	  972344	      1128 ns/op
BenchmarkGenDeleteRows
BenchmarkGenDeleteRows/pk1_ck0_col1
BenchmarkGenDeleteRows/pk1_ck0_col1-20                   	 6897980	       177.3 ns/op
BenchmarkGenDeleteRows/pk1_ck1_col1
BenchmarkGenDeleteRows/pk1_ck1_col1-20                   	 2250288	       547.2 ns/op
BenchmarkGenDeleteRows/pk3_ck3_col5
BenchmarkGenDeleteRows/pk3_ck3_col5-20                   	 1250698	       939.6 ns/op
BenchmarkGenDeleteRows/pkAll_ckAll_colAll
BenchmarkGenDeleteRows/pkAll_ckAll_colAll-20             	  343803	      3572 ns/op
BenchmarkGenDeleteRows/pk1_ck1_col1cr
BenchmarkGenDeleteRows/pk1_ck1_col1cr-20                 	 2241516	       526.9 ns/op
BenchmarkGenDeleteRows/pk3_ck3_col3cr
BenchmarkGenDeleteRows/pk3_ck3_col3cr-20                 	 1295636	       912.5 ns/op
PASS
```